// dist/node/runtime/node/community/local-community.d.ts
export declare class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalCommunityParsedOptions {
    signer: SignerWithPublicKeyAddress;
    raw: RpcLocalCommunity["raw"];
// ...
}